refactor(wupper): Add DMA descriptors enum
Add an enum that defines constants for the indexes of the DMA
descriptors. This way, the user can pass an enum member to the
dma_to_host() call instead of a literal. This approach is advised in the
method documentation. Also fixes the dma_index parameter of that method
to receive an int instead of an uint8_t.

+++ wupper_module/wupper.h
@@ -64,12 +64,13 @@ public:
      * destination. Before starting the transfer, it calls the VFIO interface to
      * map the given buffer for DMA transfers
      * @param[in] transfer_size Size in bytes of the transfer
-     * @param[in] dma_index DMA descriptor index
+     * @param[in] dma_index DMA descriptor index, expected to be a member from
+     * the enum defined in register_map
      * @param[in] buffer Reference to the DataBuffer object that points to the
      * memory buffer that will receive the transfer
      */
     void dma_to_host(
-        size_t transfer_size, uint8_t dma_index, DataBuffer &buffer);
+        size_t transfer_size, int dma_index, DataBuffer &buffer);
 
     VFIO interface;
     DataBuffer bar0;
